Create VeganEther
VeganEther is a spin-off from BitcoinVegan I own a vegan food delivery service in Charlotte. NC where I deliver vegan food to the surrounding areas, but I have recently been doing vegan coaching and selling recipes with the hopes of doing an ebook. When I first heard of Bitcoin and cryptocurrency I fell in love immediately with the control of currency, so now I want to bring veganism to the cryptocurrency world and continue to add freedom to the people with their diet. This is a great way for me to offer my services and help change the world while others help me learn as well.

Peace